+++ CHANGELOG.md
@@ -12,6 +12,18 @@
   from your shell.  
   [JP Simard](https://github.com/jpsim)
 
+* The command line syntax has slightly changed due to migrating from the
+  Commandant command line parsing library to swift-argument-parser.
+  For the most part the breaking changes are all to make the syntax more
+  unix compliant and intuitive to use. For example, commands such as
+  `sourcekitten --help` or `sourcekitten -h` now work as expected.
+  The help output from various commands has greatly improved as well.
+  Notable breaking changes are that `sourcekitten module_info` is now
+  invoked with `sourcekitten module-info`, and some commands that
+  previously took compiler arguments after a `--` delimiter should now
+  omit the `--` delimiter.  
+  [JP Simard](https://github.com/jpsim)
+
 ##### Enhancements
 
 * None.