+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ##### Enhancements
 
-* Add `key.attributes` and `key.attribute` enum cases to SwiftDocKey
+* Add `key.attributes` and `key.attribute` enum cases to SwiftDocKey.  
   [Erick Sanchez](https://github.com/ErickESGoogle)
 
 ##### Bug Fixes