+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Master
+## 0.31.1
 
 ##### Breaking